Daily precipitation grid of Balearic Islands (Spain)

Data from the SPREAD data set downloaded from the Spanish National Research Council (CSIC). (http://spread.csic.es/info.html). This data corresponds to daily values of precipitation with a spatial resolution of 5 x 5 km from January 2000 to december 2010

data(pcp)

Format

A data.frame with the following variables: lon,lat,time,value.

geographical area:

Balearic Islands

time period:

2000-01-01 to 2010-12-31

units:

mm*10

coordinates reference system:

+proj=utm +zone=30 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs

References

Serrano-Notivoli et al. (2017) SPREAD: a high-resolution daily gridded precipitation dataset for Spain, an extreme events frequency and intensity overview. Earth Syst. Sci. Data, 9, 721-738, 2017, https://doi.org/10.5194/essd-9-721-2017

Examples

data(pcp)