Self-Organizing Maps classification

som_clas allows to perform a SOM synoptic classification

Usage

som_clas(
  x,
  xdim,
  ydim,
  iter = 2000,
  alpha = c(0.05, 0.01),
  dist.fcts = "euclidean",
  mode = "online",
  cores = 1,
  norm = T
)

Arguments

x

data.frame. A data.frame with the following variables: x, y, time, value, var,units. See tidy_nc.

xdim

Integer. X dimension of the grid. See somgrid from kohonen package.

ydim

Integer. Y dimension of the grid. See somgrid from kohonen package.

iter

integer. Number of iterations.

alpha

vector. learning rate. See som from kohonen package for details.

dist.fcts

character. vector of distance functions to be used for the individual data layers. See som from kohonen package for details.

mode

carachter. type of learning algorithm. Default "on-line". See kohonen package for details.

cores

Integer. Parallel processing only available for "pbatch" algorithm.

norm

logical. Default TRUE. norm = TRUE is recommended for classifying two ore more variables.

Value

A list with:

A data.frame containing the dates and the weather types.

A data frame containing the gridded data grouped by circulation types.

An object of class kohonen with all the components returned by the function som

References

Wehrens, R. and BuydenL. (2007) Self- and Super-organizing Maps in R: The kohonen Package Journal of Statistical Software, 21(5), 1 - 19.

See also

Examples

# Load data
data(z500)

# SOM classification
som_cl <- som_clas(z500, xdim = 4, ydim = 4, iter = 200)