Objective Lamb Weather Type Classification

Calculates the classification of the main weather types for the 16-points defined in get_lamb_points. Wind-flow characteristics are computed for the daily pressure field according to the rules proposed by the original Jenkinson and Collison classification (see Jenkin-son and Collison, 1977, Jones et al. 1993, Jones et al. 2016).

lamb_clas(points, mslp)

Arguments

points
16 point pair of coordinates obtained from get_lamb_points.

mslp
Mean Sea Level pressure gridded data.

Value

A list with:

A data.frame containing the dates and the weather types.

A data frame containing the gridded data grouped by circulation types.

References

Jenkinson, A.F., Collison F.P (1977) An initial climatology of gales over the North Sea Synoptic Climatology Branch Memorandum, No. 62.Meteorological Office: Bracknell, England.Jones, P. D., Hulme M., Briffa K. R. (1993) A comparison of Lamb circulation types with an objective classification scheme Int. J. Climatol. 13: 655–663.Jones, P. D., Harpham C, Briffa K. R. (2013) Lamb weather types derived from Reanalysis products Int. J. Climatol. 33: 1129–1139.

See also

Examples

data(mslp)

points <- get_lamb_points(x = 5,y = 40)
lamb_clas(points = points, mslp = mslp)

#> $clas
#> # A tibble: 1,096 x 2
#>    time       WT   
#>    <date>     <chr>
#>  1 2000-01-01 A    
#>  2 2000-01-02 ANE  
#>  3 2000-01-03 A    
#>  4 2000-01-04 A    
#>  5 2000-01-05 A    
#>  6 2000-01-06 A    
#>  7 2000-01-07 A    
#>  8 2000-01-08 A    
#>  9 2000-01-09 C    
#> 10 2000-01-10 NE   
#> # ... with 1,086 more rows
#> 
#> $grid_clas
#> # A tibble: 5,746 x 4
#>      lon   lat WT      value
#>    <dbl> <dbl> <chr>   <dbl>
#>  1   -10    30 A     102048.
#>  2   -10    30 AE    100862.
#>  3   -10    30 AN    102502.
#>  4   -10    30 ANE   102412.
#>  5   -10    30 ANW   102522.
#>  6   -10    30 AS    101758.
#>  7   -10    30 ASE   101348.
#>  8   -10    30 ASW   101845 
#>  9   -10    30 AW    101025 
#> 10   -10    30 C     102460 
#> # ... with 5,736 more rows
#>