Objective Lamb Weather Type Classification

Calculates the classification of the main weather types for the 16-points defined in get_lamb_points. Wind-flow characteristics are computed for the daily pressure field according to the rules proposed by the original Jenkinson and Collison classification (see Jenkinson and Collison, 1977; Jones et al., 2013) (1), and to the rules proposed by Trigo and DaCamara, 2000 (2).

Usage

lamb_clas(points, msl, U = FALSE, thr = c(6, 6))

Arguments

points

16 point pair of coordinates obtained from get_lamb_points.

msl

Mean Sea Level pressure gridded data.

U

Logical. If T, Jones et al. 2013 approach is applied, maintaining the U-type in the classification. If F, U is removed as detailed in Trigo and DaCamara, 2000.

thr

threshold used for Unclassified days (total shear vorticity and total flow, respectively). Default c(6,6).

Value

A list with:

A data.frame containing the dates and the weather types.

A data frame containing the gridded data grouped by circulation types.

References

Jenkinson, A.F., Collison F.P (1977) An initial climatology of gales over the North Sea Synoptic Climatology Branch Memorandum, No. 62.Meteorological Office: Bracknell, England.Jones, P. D., Hulme M., Briffa K. R. (1993) A comparison of Lamb circulation types with an objective classification scheme Int. J. Climatol. 13: 655–663.Jones, P. D., Harpham C, Briffa K. R. (2013) Lamb weather types derived from Reanalysis products Int. J. Climatol. 33: 1129–1139.

Trigo, R., DaCamara C. (2000) Circulation weather types and their impact on the precipitation regime in Portugal Int. J. Climatol. 20: 1559-1581.

See also

Examples

data(msl)
msl$value <- msl$value/100
points <- get_lamb_points(x = 5,y = 40)
lamb_clas(points = points, msl = msl)
#> $clas
#> # A tibble: 1,096 × 2
#>    time       WT   
#>    <date>     <chr>
#>  1 2000-01-01 A    
#>  2 2000-01-02 ANE  
#>  3 2000-01-03 A    
#>  4 2000-01-04 A    
#>  5 2000-01-05 A    
#>  6 2000-01-06 A    
#>  7 2000-01-07 A    
#>  8 2000-01-08 A    
#>  9 2000-01-09 C    
#> 10 2000-01-10 NE   
#> # ℹ 1,086 more rows
#> 
#> $grid_clas
#> # A tibble: 5,746 × 4
#>        x     y WT    mean_WT_value
#>    <dbl> <dbl> <chr>         <dbl>
#>  1   -10    30 A             1018.
#>  2   -10    30 AE            1015.
#>  3   -10    30 AN            1017.
#>  4   -10    30 ANE           1015.
#>  5   -10    30 ANW           1020.
#>  6   -10    30 AS            1017.
#>  7   -10    30 ASE           1016.
#>  8   -10    30 ASW           1019.
#>  9   -10    30 AW            1019.
#> 10   -10    30 C             1016.
#> # ℹ 5,736 more rows
#>